requirements.txt
anyio==3.6.2
APScheduler==3.9.1
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==3.0.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
python-telegram-bot==20.0a0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2022.5
tzlocal==4.2
